Added stations and topics files to the applets files
Originally, these files have been fetched directly from the web.
Unfortunately, they are encoded with ISO-8859-2 which is not a correct
encoding for a XML file and lxp can't parse such files. As a workaround
the files has been added to the repository and will be loaded and parsed
from the disk directly.